[Skills Everywhere][#372] Move transcript converter out of TranscriptTestRunner (#387)
* Add TranscriptConverter command line app

* Remove TranscriptConverter from TestRunner

* Make target path optional
